refactor(core): move provider config and matchers to core
Move the provider system as a self-contained subsystem into
honeybeepf-llm-core:

  honeybeepf-llm/src/probes/builtin/llm/http/providers/{config,usage,request,mod}.rs
    -> honeybeepf-llm-core/src/providers/{config,usage,request,mod}.rs

The whole subsystem is pure logic — config structs, JSON-driven matchers,
and request text extractors — with no external dependencies beyond serde
and serde_json (already in workspace.dependencies).

Tests are moved out of the source files into core/tests/:
  - tests/providers_config.rs (3 tests)
  - tests/providers_usage.rs  (4 tests)
  - tests/providers_request.rs (4 tests, refactored to use the public
    get_extractor() factory rather than constructing private extractor
    structs directly — keeps API surface clean)

Update http/protocol.rs to import ConfigurableProvider and
ProviderRegistry from honeybeepf-llm-core::providers, and remove the
local providers module wiring from http/mod.rs.

After this PR, core has 12 tests (1 byte_utils + 11 providers) and the
main agent's lib tests drop to 6 (settings + telemetry only, both of
which depend on infrastructure and stay in the main crate).

Refs #12
