+++ honeybeepf-llm/honeybeepf-llm-core/src/lib.rs
@@ -5,4 +5,5 @@
 //! built and tested on any platform without an eBPF toolchain.
 
 pub mod byte_utils;
+pub mod providers;
 pub mod types;
