+++ honeybeepf-llm/honeybeepf-llm-core/src/providers/request.rs
@@ -84,68 +84,3 @@ impl RequestExtractor for NoOpExtractor {
         String::new()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn test_messages_extractor() {
-        let extractor = MessagesExtractor;
-        let json = json!({
-            "messages": [
-                {"role": "user", "content": "Hello"},
-                {"role": "assistant", "content": "Hi there"},
-                {"role": "user", "content": "How are you?"}
-            ]
-        });
-        let result = extractor.extract(&json);
-        assert_eq!(result, "Hello Hi there How are you?");
-    }
-
-    #[test]
-    fn test_messages_extractor_with_content_blocks() {
-        let extractor = MessagesExtractor;
-        let json = json!({
-            "messages": [
-                {
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What's in this image?"},
-                        {"type": "image_url", "image_url": {"url": "..."}}
-                    ]
-                }
-            ]
-        });
-        let result = extractor.extract(&json);
-        assert_eq!(result, "What's in this image?");
-    }
-
-    #[test]
-    fn test_contents_extractor() {
-        let extractor = ContentsExtractor;
-        let json = json!({
-            "contents": [
-                {
-                    "parts": [
-                        {"text": "Hello from Gemini"}
-                    ]
-                }
-            ]
-        });
-        let result = extractor.extract(&json);
-        assert_eq!(result, "Hello from Gemini");
-    }
-
-    #[test]
-    fn test_prompt_extractor() {
-        let extractor = PromptExtractor;
-        let json = json!({
-            "prompt": "Complete this sentence:"
-        });
-        let result = extractor.extract(&json);
-        assert_eq!(result, "Complete this sentence:");
-    }
-}
